# backend/openapi.yaml
openapi: 3.0.3
info:
  title: Reputation Management System API
  version: 1.0.0
  description: |
    API для системы управления репутацией компании. 
    Система позволяет автоматизировать сбор, анализ и мониторинг отзывов 
    о компаниях из различных источников (Яндекс, Google, 2GIS).
  contact:
    name: Reputation Management System
    email: support@reputation-system.com
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: http://localhost:8080/api
    description: Локальный сервер разработки
  - url: https://api.reputation-system.com/api
    description: Production сервер
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "JWT токен аутентификации. Получите токен через /auth/login, затем используйте его в формате: Bearer {token}"
  schemas:
...
    CreateUserRequest:
      type: object
      required:
        - username
        - email
        - password
        - fullName
        - role
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 50
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
        fullName:
          type: string
        role:
          type: string
          enum: [ADMIN, MANAGER, MODERATOR, SPECIALIST]
        companyId:
          type: integer
...
    CreateTaskRequest:
      type: object
      required:
        - reviewId
      properties:
        reviewId:
          type: integer
        assignedToId:
          type: integer
        priority:
          type: string
          enum: [LOW, MEDIUM, HIGH, URGENT]
        deadline:
          type: string
          format: date-time
    TaskResponse:
      type: object
      properties:
        id:
          type: integer
        reviewId:
          type: integer
        assignedToId:
          type: integer
        assignedToName:
          type: string
        status:
          type: string
          enum: [NEW, ASSIGNED, IN_PROGRESS, PENDING_REVIEW, IN_REVISION, COMPLETED, CANCELLED]
        priority:
          type: string
          enum: [LOW, MEDIUM, HIGH, URGENT]
        deadline:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        completedAt:
          type: string
          format: date-time
...
    CommentResponse:
      type: object
      properties:
        id:
          type: integer
        taskId:
          type: integer
        userId:
          type: integer
        userName:
          type: string
        text:
          type: string
        createdAt:
          type: string
          format: date-time
...
security:
  - bearerAuth: []
paths:
...
  /users:
    post:
      tags:
        - Управление пользователями
      summary: Создать пользователя
      description: Создание нового пользователя в системе
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
      responses:
        '201':
          description: Пользователь успешно создан
        '400':
          description: Ошибка валидации
        '403':
          description: Недостаточно прав доступа
    get:
      tags:
        - Управление пользователями
      summary: Получить всех пользователей
      description: Получение списка всех пользователей системы
      responses:
        '200':
          description: Список пользователей
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
...
  /users/{id}:
    get:
      tags:
        - Управление пользователями
      summary: Получить пользователя по ID
      description: Получение информации о пользователе по идентификатору
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Информация о пользователе
        '404':
          description: Пользователь не найден
  /users/{id}/reset-password:
    put:
      tags:
        - Управление пользователями
      summary: Сбросить пароль пользователя
      description: Сброс пароля пользователя администратором
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Пароль успешно сброшен
        '404':
          description: Пользователь не найден
  /users/me/password:
    put:
      tags:
        - Управление пользователями
      summary: Изменить свой пароль
      description: Изменение пароля текущего пользователя
      responses:
        '200':
          description: Пароль успешно изменен
        '400':
          description: Ошибка валидации
...
  /tasks:
    post:
      tags:
        - Управление задачами
      summary: Создать задачу
      description: Создание новой задачи для работы с отзывом
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTaskRequest'
      responses:
        '201':
          description: Задача успешно создана
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
    get:
      tags:
        - Управление задачами
      summary: Получить все задачи
      description: Получение списка всех задач
      responses:
        '200':
          description: Список задач
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TaskResponse'
  /tasks/{id}:
    get:
      tags:
        - Управление задачами
      summary: Получить задачу по ID
      description: Получение информации о задаче по идентификатору
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Информация о задаче
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '404':
          description: Задача не найдена
...
  /tasks/{id}/assign:
    put:
      tags:
        - Управление задачами
      summary: Назначить задачу
      description: Назначение задачи специалисту
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: userId
          in: query
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Задача успешно назначена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '404':
          description: Задача или пользователь не найдены
  /tasks/{id}/start:
    put:
      tags:
        - Управление задачами
      summary: Начать работу над задачей
      description: Начало работы специалиста над задачей
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Работа над задачей начата
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '404':
          description: Задача не найдена
  /tasks/{id}/complete:
    put:
      tags:
        - Управление задачами
      summary: Завершить задачу
      description: Завершение работы над задачей и отправка на проверку
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Задача завершена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '404':
          description: Задача не найдена
  /tasks/{id}/cancel:
    put:
      tags:
        - Управление задачами
      summary: Отменить задачу
      description: Отмена задачи
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Задача отменена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '404':
          description: Задача не найдена
  /tasks/{id}/return:
    put:
      tags:
        - Управление задачами
      summary: Вернуть задачу на доработку
      description: Возврат задачи на доработку специалисту
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Задача возвращена на доработку
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '404':
          description: Задача не найдена
...
  /comments/task/{taskId}:
    get:
      tags:
        - Работа с комментариями
      summary: Получить комментарии к задаче
      description: Получение списка комментариев для конкретной задачи
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Список комментариев
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CommentResponse'
...